message-generator/requirements.txt
kafka-python
confluent-kafka[avro]
pandas
seaborn